Map floating point operations to c-calls instead of intrinsics.
Originally the point was to implement floating point with a library of fixed point
arithmetic routines.  It also added a fixpt environment parameter and generated
fixpt constants.

We drop that and just do the mapping to c-calls.  The hope is we can then map
floating point to softfloat with clang.  This avoids a great deal of changes
to the caml runtime for the kernel module.  If it works.